Added Threshold and Notification Settings for Power Meter devices
Use Case: I have a Z-Wave outlet that reports power usage in Watts.  I need to be notified if the power usage is outside its normal threshold.  In my case, it's for a pump for a pond with fish in it.  If the pump dies, the power usage will (presumably) go down to somewhere near zero, even though the outlet will still be on and operational.  If this goes on for too long without my being aware, the fish can die.  Receiving a notification when the power usage in Watts drops below a certain threshold can help to keep my fish alive.

I've tested this for too-high, too-low, and within the threshold notifications.  I modeled it after the code for temperature  monitoring.

I tried to follow the patterns and conventions I saw, but please let me know if anything needs to be changed, and I'll be glad to update the pull request.